chore: major dependency updates with numpy 2 migration
Major version bumps:
- numpy 1.26 → 2.4 (added explicit bool/float/int casts for numpy 2)
- torch 2.7 → 2.11, torchvision 0.22 → 0.26
- transformers 4.46 → 4.57, tokenizers 0.20 → 0.22
- black 24 → 26, ruff 0.14 → 0.15 (migrated str,Enum to StrEnum)
- pytest 8 → 9, pytest-asyncio 0.24 → 1.3, pytest-cov 5 → 7
- sphinx 7 → 9, myst-parser 4 → 5, pylint 3 → 4
- opencv 4.11 → 4.13, pillow 11 → 12, mss 9 → 10
- croniter 5 → 6, prometheus-client 0.19 → 0.24
- supervision 0.26 → 0.27

Minor/patch: requests, pyyaml, anyio, types-requests, timm, tox,
python-discovery, faiss-cpu, ultralytics, sphinx-autodoc-typehints,
sphinx-rtd-theme

+++ src/qontinui/perception/vectorization.py
@@ -167,11 +167,11 @@ class ObjectVectorizer:
         v_projection = np.mean(edges, axis=0)
 
         features = [
-            edge_density,
-            np.mean(h_projection),
-            np.std(h_projection),
-            np.mean(v_projection),
-            np.std(v_projection),
+            float(edge_density),
+            float(np.mean(h_projection)),
+            float(np.std(h_projection)),
+            float(np.mean(v_projection)),
+            float(np.std(v_projection)),
         ]
 
         return features
